Improvements to Logging and Update Script
Improved Log-Message function and added a "Skip" level. Cleaned up and clarified console output throughout script. Improved update filtering in update script along with improving console legibility. Remove simulated sleeps to speed up script progress.

+++ PC Setup Script.ps1
@@ -8,6 +8,20 @@ if (-not $IsElevated) {
 }
 
 # Script setup
+try {
+	$dWidth = (Get-Host).UI.RawUI.BufferSize.Width
+	$dHeight = 40
+	$rawUI = $Host.UI.RawUI
+	$newSize = New-Object System.Management.Automation.Host.Size ($dWidth, $dHeight)
+	$rawUI.WindowSize = $newSize
+} catch {
+	$failedResize = 1
+}
+try {
+	$host.UI.RawUI.BackgroundColor = "Black"
+} catch {
+	$failedColor = 1
+}
 Clear-Host
 $DesktopPath = [Environment]::GetFolderPath('Desktop')
 $logPathName = "PCSetupScriptLog.txt"
@@ -16,6 +30,8 @@ $WUSPath = Join-Path -Path $PSScriptRoot -ChildPath 'Windows Update Script.ps1'
 $functionPath = Join-Path -Path $PSScriptRoot -ChildPath 'Script Functions.ps1'
 . "$functionPath"
 $serialNumber = (Get-WmiObject -Class Win32_BIOS).SerialNumber
+if ($failedResize -eq 1) {Log-Message "Failed to resize window." "Error"}
+if ($failedColor -eq 1) {Log-Message "Failed to change background color." "Error"}
 
 # Set time zone and sync
 Log-Message "Setting Time Zone to Eastern Standard Time..."
@@ -37,42 +53,47 @@ Start-Process -FilePath "powershell.exe" -ArgumentList "-ExecutionPolicy Bypass"
 Log-Message "Setup Local Account(s)..."
 $RepeatFunction = 1
 While ($RepeatFunction -eq 1) {
-    Log-Message "Please enter a username:" "Prompt"
+    Log-Message "Please enter a username or leave blank to skip this section:" "Prompt"
 	$AdminUser = Read-Host
-    Log-Message "Please enter a password or leave blank to skip:" "Prompt"
-	$AdminPass = Read-Host
-	$UExists = Get-LocalUser -Name $AdminUser -ErrorAction SilentlyContinue
-	if (-not $UExists) {
-		Log-Message "The specified user does not exist, create account now? (y/N)" "Prompt"
-        $MakeUser = Read-Host
-		if ($MakeUser -eq "y" -or $MakeUser -eq "Y") {
-			Net User $AdminUser $AdminPass /add | Out-File -Append -FilePath $logPath
+	if (-not $AdminUser -eq "") {
+		Log-Message "Please enter a password (can be empty):" "Prompt"
+		$AdminPass = Read-Host
+		$UExists = Get-LocalUser -Name $AdminUser -ErrorAction SilentlyContinue
+		if (-not $UExists) {
+			Log-Message "The specified user does not exist, create account now? (y/N):" "Prompt"
+			$MakeUser = Read-Host
+			if ($MakeUser -eq "y" -or $MakeUser -eq "Y") {
+				Net User $AdminUser $AdminPass /add | Out-File -Append -FilePath $logPath
+			} else {
+				Log-Message "Skipping account creation." "Skip"
+			}
 		} else {
-			Log-Message "Skipping account creation."
+			Log-Message "Update the user's password? (y/N):" "Prompt"
+			$UpdateUser = Read-Host
+			if ($UpdateUser.ToLower() -eq "y" -or $UpdateUser.ToLower() -eq "yes") {
+				Net User $AdminUser $AdminPass | Out-File -Append -FilePath $logPath
+			}
+		}
+		$LocalUserCheck = "$env:COMPUTERNAME\$AdminUser"
+		$IsAdmin = Get-LocalGroupMember -Group "Administrators" -ErrorAction SilentlyContinue | Where-Object { $_.Name -eq $LocalUserCheck }
+		if ($UExists -and -not $IsAdmin) {
+			Log-Message "The specified user is not a local admin, elevate now? (y/N):" "Prompt"
+			$MakeAdmin = Read-Host
+			if ($MakeAdmin -eq "y" -or $MakeAdmin -eq "Y") {
+				Net Localgroup Administrators $AdminUser /add | Out-File -Append -FilePath $logPath
+			} else {
+				Log-Message "Skipping account elevation." "Skip"
+			}
+		} elseif ($UExists -and $IsAdmin) {
+			Log-Message "Skipping account elevation, user account is already a local administrator." "Skip"
+		}
+		Log-Message "Repeat this segment to add, edit or test another user account? (y/N):" "Prompt"
+		$RFQ = Read-Host
+		if (-not ($RFQ.ToLower() -eq "y" -or $RFQ.ToLower() -eq "yes")) {
+			$RepeatFunction = 0
 		}
 	} else {
-        Log-Message "Update the user's password? (y/N)" "Prompt"
-		$UpdateUser = Read-Host
-		if ($UpdateUser.ToLower() -eq "y" -or $UpdateUser.ToLower() -eq "yes") {
-			Net User $AdminUser $AdminPass | Out-File -Append -FilePath $logPath
-		}
-	}
-	$LocalUserCheck = "$env:COMPUTERNAME\$AdminUser"
-	$IsAdmin = Get-LocalGroupMember -Group "Administrators" -ErrorAction SilentlyContinue | Where-Object { $_.Name -eq $LocalUserCheck }
-	if ($UExists -and -not $IsAdmin) {
-        Log-Message "The specified user is not a local admin, elevate now? (y/N)" "Prompt"
-		$MakeAdmin = Read-Host
-		if ($MakeAdmin -eq "y" -or $MakeAdmin -eq "Y") {
-			Net Localgroup Administrators $AdminUser /add | Out-File -Append -FilePath $logPath
-		} else {
-			Log-Message "Skipping account elevation."
-		}
-	} elseif ($UExists -and $IsAdmin) {
-		Log-Message "Skipping account elevation, user account is already a local administrator."
-	}
-    Log-Message "Repeat this segment to add, edit or test another user account? (y/N)" "Prompt"
-	$RFQ = Read-Host
-	if (-not ($RFQ.ToLower() -eq "y" -or $RFQ.ToLower() -eq "yes")) {
+		Log-Message "Skipping account management." "Skip"
 		$RepeatFunction = 0
 	}
 }
@@ -85,22 +106,24 @@ Log-Message "Updating System Packages and Apps (This may take some time)..."
 WinGet Upgrade --ALL --scope machine --accept-package-agreements --accept-source-agreements | Out-File -Append -FilePath $logPath
 
 # Remove commond Windows bloat
-Log-Message "Would you like to remove common Windows bloat programs? (y/N)" "Prompt"
+Log-Message "Would you like to remove common Windows bloat programs? (y/N):" "Prompt"
 $RemoveBloat = Read-Host
 if ($RemoveBloat.ToLower() -eq "y" -or $RemoveBloat.ToLower() -eq "yes") {
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*bingfinance*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*bingnews*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*bingsports*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*gethelp*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*getstarted*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*mixedreality*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*people*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*solitaire*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*wallet*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*windowsfeedback*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*windowsmaps*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*xbox*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
-	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*zunevideo*" | Remove-AppxPackage -AllUsers -Verbose | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*bingfinance*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*bingnews*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*bingsports*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*gethelp*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*getstarted*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*mixedreality*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*people*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*solitaire*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*wallet*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*windowsfeedback*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*windowsmaps*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*xbox*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+	Get-AppxPackage -AllUsers -PackageTypeFilter Bundle -Name "*zunevideo*" | Remove-AppxPackage -AllUsers -Verbose 4>&1 | Out-File -Append -FilePath $logPath
+} else {
+	Log-Message "Skipping bloat removal." "Skip"
 }
 
 # Install programs based on selections, prepare Windows "Form"
@@ -204,7 +227,7 @@ $okButton.Add_Click({
     $selectedPrograms = $checkboxes.GetEnumerator() | Where-Object { $_.Value.Checked } | ForEach-Object { $_.Key }
     $totalPrograms = $selectedPrograms.Count
     if ($totalPrograms -eq 0) {
-        Log-Message "No programs selected for installation. Exiting." -ForegroundColor Yellow
+        Log-Message "No programs selected for installation." "Skip"
         $form.Close()
         return
     }
@@ -235,14 +258,14 @@ $okButton.Add_Click({
                 # Capture the result
                 if ($process.ExitCode -eq 0) {
                     $message = "$($program.Name): Installed successfully."
-                    Log-Message $message -ForegroundColor Green
+                    Log-Message $message "Success"
                 } else {
                     $message = "$($program.Name): Installation failed with exit code $($process.ExitCode)."
-                    Log-Message $message -ForegroundColor Red
+                    Log-Message $message "Error"
                 }
             } catch {
                 $message = "$($program.Name): Installation failed. Error: $_"
-                Log-Message $message -ForegroundColor Red
+                Log-Message $message "Error"
             }
         }
         $progressBar.Value += 1
@@ -258,16 +281,16 @@ $form.ShowDialog() | Out-null
 
 # Rename PC and join to domain (if needed)
 Log-Message "The PC is currently named: $env:computername"
-Log-Message "Would you like to change the PC name? y/n" "Prompt"
+Log-Message "Would you like to change the PC name? (y/N):" "Prompt"
 $Rename = Read-Host
 if ($Rename -eq "y" -or $Rename -eq "Y") {
     Log-Message "The serial number is: $serialNumber"
-    Log-Message "Enter the new PC name and press Enter" "Prompt"
+    Log-Message "Enter the new PC name and press Enter:" "Prompt"
     $PCName = Read-Host
-    Log-Message "Would you like to join this PC to an Active Directory Domain? y/n" "Prompt"
+    Log-Message "Would you like to join this PC to an Active Directory Domain? (y/N):" "Prompt"
 	$Domain = Read-Host
 	if ($Domain -eq "y" -or $Domain -eq "Y") {
-        Log-Message "Enter the domain address and press Enter (Include the suffix, Ex: .local)" "Prompt"
+        Log-Message "Enter the domain address and press Enter (Include the suffix, Ex: .local):" "Prompt"
 		$DomainName = Read-Host
 		$DomainCredential = Get-Credential -Message "Enter credentials with permission to add this device to $DomainName"
 		Add-Computer -DomainName $DomainName -NewName $PCName -Credential $DomainCredential | Out-File -Append -FilePath $logPath
@@ -275,15 +298,17 @@ if ($Rename -eq "y" -or $Rename -eq "Y") {
 		Rename-Computer -NewName $PCName -Force | Out-File -Append -FilePath $logPath
 	}
 } else {
-    Log-Message "Would you like to join this PC to an Active Directory Domain? y/n" "Prompt"
+    Log-Message "Would you like to join this PC to an Active Directory Domain? (y/N):" "Prompt"
 	$Domain = Read-Host
 	if ($Domain -eq "y" -or $Domain -eq "Y") {
-        Log-Message "Enter the domain address and press Enter (Include the suffix, Ex: .local)" "Prompt"
+        Log-Message "Enter the domain address and press Enter (Include the suffix, Ex: .local):" "Prompt"
 		$DomainName = Read-Host
-		$DomainCredential = Get-Credential -Message "Enter credentials with permission to add this device to $DomainName"
+		$DomainCredential = Get-Credential -Message "Enter credentials with permission to add this device to $DomainName:"
 		Add-Computer -DomainName $DomainName -Credential $DomainCredential | Out-File -Append -FilePath $logPath
 	}
 }
+if ($Domain -eq "y" -or $Domain -eq "Y" -or $Rename -eq "y" -or $Rename -eq "Y") {
+	Log-Message "PC rename and/or domain join complete." "Success"
 
 # Final setup options
 $regPathNumLock = "Registry::HKEY_USERS\.DEFAULT\Control Panel\Keyboard"
@@ -297,8 +322,7 @@ if (Test-Path $regPathNumLock) {
 
 # Reminders/Closing
 Log-Message "Script setup is complete!"
-Log-Message "Please install the agent and make any remaining changes needed."
-Log-Message "Confirm Windows Updates have completed in the minimzied window and restart if needed."
+Log-Message "Confirm updates have completed in the minimized window and restart to apply updates, PC name change and domain joining if needed."
 Log-Message "Press enter to exit the script." "Success"
 Read-Host
 
@@ -306,4 +330,4 @@ Read-Host
 $folderToDelete = $PSScriptRoot
 $deletionCommand = "Start-Sleep -Seconds 2; Remove-Item -Path `"$folderToDelete`" -Recurse -Force"
 Start-Process powershell.exe -ArgumentList "-NoProfile", "-WindowStyle", "Hidden", "-Command", $deletionCommand
-exit
+exit 0
